Add new module-list settings that take globs instead of regexps
The list of regexps are not very readable. But we want to keep this
syntax for backward compatibility. So add a set of new settings that
take globs instead:
KernelModulesIncludeGlob=
KernelModulesExcludeGlob=
KernelModulesInitrdIncludeGlob=
KernelModulesInitrdExcludeGlob=

The globs are interpreted in a way that makes the most sense for
matching modules and module subhierarchies:
- 'foo' matches the basename, /some/path/foo.ko
- 'bar/foo' matches the last component of the path, bar/foo.ko
- '/full/path/bar' only matches /full/path/bar.ko.
- crypto/* matches all modules in crypto subdirs
- /crypto/* matches the modules in the top-level subdir

Suffixes are not specified in the globs. '-' and '_' are treated
as equivalent, except when part of special glob syntax with […].

I think this is both more pleasant to use and more readable.