debian: use libeatmydata with dpkg
Even with --force-unsafe-io dpkg still does a lot of fsyncs.
Pull libeatmydata1 together with the bootstrapping set, and
LD_PRELOAD it. It makes all fsync and related syscalls no-ops.

Before: 2m23.104s

After: 1m8.610s

The package is ~8KB and installed is ~30KB so it should make no
difference to have it installed, and it's inert unless preloaded.