setup: use setuptools-scm to generate the version
This commit changes the versioning scheme of non-release versions of mkosi to
align with the standard Python version scheme [1]. Unfortunately this scheme is
not fully compatible with the UAPI group version format specification.

Concretely, this changes

    26~devel

to

    25.3.post1.dev244+g0eac3718

The used format ensures that for prereleases the following inequalities still
hold

    25.3 < 25.3.post1.dev244+g0eac3718 < 25.4 < 26

Usage of this new format has the following benefits:

1. It is compatible with PyPI and allows publishing mkosi there in the future

2. Since it contains the number of commits since the last tag as well as part
   of the commit hash in the "local part" of the version, it becomes easier for
   bug reporters to show the version they are on no matter how mkosi has been
   installed.

3. The version no longer needs to be managed in multiple places.

The current implementation moves the version from mkosi/sandbox.py to a new
file _version.py that obtains the version either via importlib metadata (in
case mkosi has been installed) or by calling setuptools-scm directly, to
support the bin/ shims. Since the version is no longer available in
mkosi/sandbox.py, it passed as an environment variable for scripts.

[1] https://packaging.python.org/en/latest/specifications/version-specifiers/