qemu: Make notify handling code more generic
Let's pass in a queue to the target function of AsyncioThread() to which
it can write messages. Let's also keep track of all sent messages in
AsyncioThread for easy access later and add some utility functions to
process new messages.

We also split out notify() from vsock_notify_handler() and make it non-vsock
specific by passing in the socket as an argument.