action: Ensure git history is available
Unshallowing the repository only works when the action is used as a
local action within the mkosi repository. If the action is used in any
other github repository, the git history is not available so we cannot
unshallow the repository.

Instead, detect if the git history is available and clone the repository
ourselves if the git history is not available.