detect kernel ver from modules sub dir if unable to extract from filename
Some distros don't include the kernel version in the filename of the
kernel image, so the normal version detection mechanisms don't work.
Grabbing the first dir under this path that looks like a kernel version
should work in these cases, since mkosi is always building a new image
there should only ever be 1 modules dir and it should be for the only
kernel that was installed.