run.py: always log the sandbox when a command fails
The main focus of commit 7e7a3c71696d6 ("Don't log sandbox for every
command") was to reduce verbosity in the --debug case and it did just
that. To avoid losing the sandbox completely, that commit said "... but
still log the full sandbox if a command fails" and it did that too in
__init__.py; for both the --debug and non-debug cases.

However, what happened in run.py was a bit different: there, commit
7e7a3c71696d6 showed the sandox on failure but only in the --debug
case!? Make run.py consistent with __init__.py and always show the
sandbox on failure in run.py, in _both_ --debug and non-debug case.

This helps with issue #3948: hiding that bind mounts exist is especially
confusing on fatal "disk full" failures because that leaves only
misleading paths in the error message! On failure, showing such critical
sandbox information must not require --debug.

Not concealing the sandbox on failure is useful in any case, not
just #3948

Signed-off-by: Marc Herbert <marc.herbert@intel.com>