Add KernelModules needed by stock fedora to initrd[fedora.conf]
zram-generator has been on by default in fedora
since fc33 (2020) and runs during early boot.
If the zram module is missing, users are hit with a nasty
45s systemd timeout.

firewalld.service is also on by default and fails to start
without nfnetlink.

Let's make sure an image built with `KernelModules=default`,
which pulls in the default mkosi-initrd modules, brings up
a fedora system in a non-degraded state

Before:
52038188 fedora.initrd

After:
52105776 fedora.initrd