Don't discard ordering of include/exclude entries in KernelModules=
Since eecf8b3b5 `KernelModules=` supports both inclusion and exclusion
of kmods. When resolving kmod presets like "default" and "host"
(and perhaps others in the future) it's important to retain the
ordering of entries, so that the user can add a preset and then
subtract only specific modules from the working set.

`filter_kernel_modules` relies on this ordering to do its job.