initd: always compress kmod initrd
Compress with level '-0' on newer distributions where modules are
individually compressed.
Concatenating compressed and uncompressed cpio archives in the same
file makes it impossible to extract them again to inspect the
content, as zstd chokes when a zstd and non-zstd content is in the
same file, and cpio chokes when the opposite happens.

By using -0 and always having zstd content, one can just pipe
unzstd into cpio and it works fine.